* update270: (22 commits)
  chore: release v2.7.0
  fix: allow `0` as an animation duration on both source and thumbnail
  chore: updated CHANGELOG.md
  chore: release v2.6.0
  chore: eslint fixes
  chore: updated example packages, react-native@0.72.3
  chore: updated dev packages, react-native@0.72.3
  fix: check if path exist when clearing
  chore: release v2.5.0
  chore: release v2.5.0
  chore: updated dev end example packages
  fix for "Image source "null"" warning
  chore: release v2.4.0
  chore: updated eslint rules
  chore: prettified files
  chore: added missing / on bugs.url on package.json
  chore: added nullish check on sourceAnimationDuration from global config
  chore: added tintColor prop on source image
  chore: added missing comma on package.json
  chore: updated example to react-native@0.71.8 and dev packages
  ...

+++ src/CacheManager.ts
@@ -103,7 +103,11 @@ export default class CacheManager {
     noCache?: boolean,
     maxAge?: number
   ): CacheEntry {
-    if (!CacheManager.entries[source]) {
+    if (
+      !CacheManager.entries[source] ||
+      CacheManager.entries[source].options?.headers?.Authorization !==
+        options?.headers?.Authorization
+    ) {
       CacheManager.entries[source] = new CacheEntry(
         source,
         options,
@@ -116,12 +120,14 @@ export default class CacheManager {
   }
 
   static async clearCache(): Promise<void> {
-    const files = await FileSystem.ls(CacheManager.config.baseDir);
-    for (const file of files) {
-      try {
-        await FileSystem.unlink(`${CacheManager.config.baseDir}${file}`);
-      } catch (e) {
-        console.log(`error while clearing images cache, error: ${e}`);
+    if (await FileSystem.exists(CacheManager.config.baseDir)) {
+      const files = await FileSystem.ls(CacheManager.config.baseDir);
+      for (const file of files) {
+        try {
+          await FileSystem.unlink(`${CacheManager.config.baseDir}${file}`);
+        } catch (e) {
+          console.log(`error while clearing images cache, error: ${e}`);
+        }
       }
     }
   }
